+++ Team Unknown/src/enemy/Enemy1.java
@@ -9,7 +9,7 @@ public class Enemy1 extends Enemy{
 	}
 
 	@Override
-	public void doDamage(int damage) {
+	public void doDamage(double damage) {
 		health -= damage;
 		System.out.println("Enemy health: "+health);
 	}
